Node breakup: port mapper
